Parse long to avoid NumberFormatException on large numbers
e.g. for https://nwbib.de/990054518830206441

See also https://github.com/hbz/lobid-resources/pull/1807

+++ app/controllers/nwbib/Lobid.java
@@ -808,8 +808,8 @@ public class Lobid {
 		return String.format("http://lobid.org/resources/%s#!", shortId);
 	}
 
-	private static int numerical(String s) {
+	private static long numerical(String s) {
 		// replace non-digits with 9, e.g. for DE-5 before DE-Walb1
-		return Integer.parseInt(s.replaceAll("\\D", "9"));
+		return Long.parseLong(s.replaceAll("\\D", "9"));
 	}
 }
